S1.01.04：增加FOC驱动框架
feature(spi_bus):1.增加硬件spi的xfer接口。
feature(foc):1.将除法改为乘以倒数。2.仅调试用，dq给定值。
feature(tle5012b：1.优化crc性能，从计算改为查表。

+++ motion/foc/math/foc_math_backend.h
@@ -22,7 +22,7 @@
 
 /*---------- type define ----------*/
 typedef float foc_scalar_t;
-typedef uint16_t foc_angle_t;
+typedef float foc_angle_t;
 
 typedef struct {
     /* 基于当前后端计算得到的正余弦值 */
